requirements.txt
flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.0
openai==1.12.0
python-dotenv==1.0.0
